resolving unique IDs in the Sqlite and starting the CheckDescriptionScreen
