+++ MedicineReminder/Source/Feature/CheckPrescriptions/View/CheckPrescriptionsViewProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  CheckPrescriptionsViewProtocol.swift
+//  MedicineReminder
+//
+//  Created by Mateus Henrique Coelho de Paulo on 06/02/22.
+//
+
+import Foundation
+
+protocol CheckPrescriptionsViewProtocol where self: UIView {
+    
+}
